+++ CHANGELOG.md
@@ -15,6 +15,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 
+## [0.7.0] - 2021-08-27
+
+### Added
+
+* Added custom swipe implementation based on touch events (#207, #208)
+* Added image caption template system (#6, #154)  
+* Added Angular material to the demo page (#212)
+* Added configuration for caption colors (#213, #212)
+
+### Changed
+
+* Fixed issues with pointer on button (#212)
+* Fixed issue with overlay and angular material (z-index issue) (#212)
+
+### Removed
+
+* Removed `hammerjs` as a peer dependency
+
 ## [0.6.1] - 2021-06-21
 
 ### Added